\name{exampleone}
\alias{exampleone}

\docType{data}
\title{
Leading causes of death, total population, by age group
}
\description{
Leading causes of death, total population, by age group at time of death, 1 to 14 years 
}
\usage{data("exampleone")}
\format{
\code{exampleone} is a data set, with columns being the year
 and row being the leading causes of death.
}

\source{ 
\url{https://doi.org/10.25318/1310039401-eng}
}

\references{
Statistics Canada.  Table  13-10-0394-01   Leading causes of death, total population, by age group
}

\keyword{datasets}




