\name{exampletwo}
\alias{exampletwo}
\docType{data}
\title{
Leading causes of death, total population, by age group
}
\description{
Leading causes of death, total population, by age group at time of death, all ages
}
\usage{data("exampletwo")}
\format{
  A data frame with 51 observations on the 19 variables, with columns being the year
 and row being the leading causes of death.
  }


\source{ 
\url{https://doi.org/10.25318/1310039401-eng}
}

\references{
Statistics Canada.  Table  13-10-0394-01   Leading causes of death, total population, by age group
}
\examples{
data(exampletwo)
## maybe str(exampletwo) ; plot(exampletwo) ...
}
\keyword{datasets}
