% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_sim.R
\name{fisher.sim}
\alias{fisher.sim}
\title{Fisher's exact test on GPU}
\usage{
fisher.sim(
  x,
  C,
  streams,
  type = c("float", "double")[1 + gpuR::gpuInfo()$double_support],
  returnStatistics = FALSE,
  Nglobal,
  verbose = FALSE
)
}
\description{
Fisher's exact test on GPU
}
