% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuRn.R
\name{gpuRn}
\alias{gpuRn}
\title{Random number generation on GPU}
\usage{
gpuRn(x, streams, workgroupSize, localSize = c(2, 2),
  random_type = c("uniform", "normal"))
}
\arguments{
\item{x}{A Vclvector of returned random numbers}

\item{streams}{Vectors of integers, 18x workgroupSize}

\item{workgroupSize}{Global dimensions of work-item domain}

\item{localSize}{Dimensions of work groups}

\item{random_type}{Uniform or Normal random numbers}
}
\value{
Altered in-place
}
\description{
Generate uniform or Normal random numbers on GPU.
}
