% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_sim.R
\name{fisher.sim}
\alias{fisher.sim}
\title{Fisher's exact test on GPU}
\usage{
fisher.sim(
  x,
  N,
  streams,
  type = c("float", "double")[1 + gpuR::gpuInfo()$double_support],
  returnStatistics = FALSE,
  Nglobal,
  verbose = FALSE
)
}
\arguments{
\item{x}{a vclMatrix for Fisher's test}

\item{N}{requested number of simulations.}

\item{streams}{streams objects.}

\item{type}{"double" or "float" of test statistics.}

\item{Nglobal}{the size of the index space for use.}
}
\value{
a list of results
}
\description{
Fisher's exact test on GPU
}
