% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnorm.R
\name{qqnorm}
\alias{qqnorm}
\title{qqnorm on GPU}
\usage{
qqnorm(
  y,
  ylim,
  mu = 0,
  sigma = 1,
  lowertail = 1,
  main = "Normal Q-Q Plot",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  Nglobal,
  Nlocal = c(2, 2),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{y}{the data sample}

\item{ylim}{limits on the plot region}

\item{mu}{mean of Normal distribution}

\item{sigma}{variance of Normal distribution}

\item{lowertail}{logical, whether use lower tail probability, default is TRUE}

\item{main}{plot label}

\item{xlab}{plot label}

\item{ylab}{plot label}

\item{Nglobal}{the size of the index space for use, default is 64 by 4}

\item{Nlocal}{the work group size of the index space}
}
\value{
the Normal Q-Q plot

the result matrices of LDB in batch
}
\description{
qqnorm on GPU
}
