% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnorm.R
\name{qqnorm}
\alias{qqnorm}
\title{gpu_qqnorm}
\usage{
qqnorm(
  y,
  ylim,
  mu = 0,
  sigma = 1,
  lowertail = 1,
  main = "Normal Q-Q Plot",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  Nglobal,
  Nlocal = c(2, 2),
  verbose = FALSE,
  ...
)
}
\description{
gpu_qqnorm
}
