% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm.R
\name{rnorm}
\alias{rnorm}
\title{rnorm}
\usage{
rnorm(n, streams, Nglobal, type = c("double", "float"))
}
\arguments{
\item{n}{a number or a vector specifying the size of output matrix}

\item{streams}{streams matrix of random number seeds}

\item{Nglobal}{vector of length 2, number of work items}

\item{type}{precision type of random numbers, "double" or "float"}
}
\value{
a vclVector or vclMatrix of standard normal random numbers
}
\description{
rnorm
}
