% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif.R
\name{runif}
\alias{runif}
\title{runif}
\usage{
runif(n, streams, Nglobal, type = c("double", "float"))
}
\arguments{
\item{n}{a number or a vector specifying the size of output matrix}

\item{streams}{streams object}

\item{Nglobal}{the size of the index space for use}

\item{type}{"double" or "float" format of generated random numbers}
}
\value{
a vclVector or vclMatrix of uniform random numbers
}
\description{
runif
}
