% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif.R
\name{runif}
\alias{runif}
\title{runif}
\usage{
runif(n, streams, Nglobal, type = c("double", "float", "integer"))
}
\arguments{
\item{n}{a number or a vector specifying the size of output matrix}

\item{streams}{streams matrix of random number seeds}

\item{Nglobal}{vector of length 2, number of work items}

\item{type}{precision type of random numbers, "double" or "float" or "integer"}
}
\value{
a vclVector or vclMatrix of uniform random numbers
}
\description{
runif
}
